move test to prose test file
